+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ## Changes
 - Added support for URLs in the `Room Code` input
+- Added condition `has_obtained_needle` with 4 simple goals
 
 </details>